Change tools and tags list and refactor code
This contains the code adjustments for #368. The lists of tools and tags were flattened, which makes rendering the data into different formats easier. This is especially beneficial for the website.

Some notes about changes from the refactoring:

Switched the templating engine from Tera to Askama due to Keats/tera#522. The alternative would have been to restructure the code, but the current structure felt more natural and more in line with what we plan for the future.
I noticed that some tags did not contain any tools, so that's why they were removed from the list.